+++ src/main/java/com/melek/gestionstock/service/impl/FournisseurServiceImpl.java
@@ -3,6 +3,11 @@ package com.melek.gestionstock.service.impl;
 import com.melek.gestionstock.dto.FournisseurDto;
 import com.melek.gestionstock.exception.EntityNotFoundException;
 import com.melek.gestionstock.exception.ErrorCodes;
+import com.melek.gestionstock.exception.InvalidOperationException;
+import com.melek.gestionstock.model.CommandeFournisseur;
+import com.melek.gestionstock.model.LigneCommandeClient;
+import com.melek.gestionstock.model.LigneCommandeFournisseur;
+import com.melek.gestionstock.repository.CommandeFournisseurRepository;
 import com.melek.gestionstock.repository.FournisseurRepository;
 import com.melek.gestionstock.service.IFournisseurService;
 import com.melek.gestionstock.validator.FournisseurValidator;
@@ -18,10 +23,12 @@ import java.util.stream.Collectors;
 public class FournisseurServiceImpl implements IFournisseurService {
 
     private FournisseurRepository fournisseurRepository;
+    private CommandeFournisseurRepository commandeFournisseurRepository;
 
     @Autowired
-    public FournisseurServiceImpl(FournisseurRepository fournisseurRepository) {
+    public FournisseurServiceImpl(FournisseurRepository fournisseurRepository, CommandeFournisseurRepository commandeFournisseurRepository) {
         this.fournisseurRepository = fournisseurRepository;
+        this.commandeFournisseurRepository = commandeFournisseurRepository;
     }
 
     @Override
@@ -58,6 +65,10 @@ public class FournisseurServiceImpl implements IFournisseurService {
     public void delete(Integer id) {
         if (id == null) {
             log.warn("Id is null");
+        }
+        List<CommandeFournisseur> commandeFournisseurs = commandeFournisseurRepository.findAllByFournisseurId(id);
+        if (!commandeFournisseurs.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une commande fournisseur déjà utilisé", ErrorCodes.FOURNISSEUR_ALREADY_IN_USE);
         }
         fournisseurRepository.deleteById(id);
     }
